requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiogram-calendar==0.5.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
et-xmlfile==1.1.0
frozenlist==1.4.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==4.6.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
